Fix remaining compile errors: ffmpeg-kit Statistics.time type, bad import
- AudioVideoConverter.kt: Statistics.time is Double in this ffmpeg-kit
  fork (not Long as assumed), causing a Float/Double mismatch against
  coerceIn(0f, 0.99f) and the Float-typed onProgress callback.
- ConverterScreen.kt: ExposedDropdownMenu is a member function of
  ExposedDropdownMenuBoxScope, not a top-level composable — importing it
  from androidx.compose.material3 doesn't resolve. The call site inside
  the ExposedDropdownMenuBox content lambda already resolves it via the
  implicit receiver, so the import was both wrong and unnecessary.

All three build errors reported by an actual Android Studio compile
(this sandbox can't run one) are fixed.

+++ app/src/main/java/com/megaconverter/app/converter/converters/AudioVideoConverter.kt
@@ -64,7 +64,7 @@ class AudioVideoConverter : FileConverter {
                 { /* ignore raw ffmpeg logs */ },
                 { statistics ->
                     if (durationMs > 0) {
-                        val progress = (statistics.time / durationMs.toFloat()).coerceIn(0f, 0.99f)
+                        val progress = (statistics.time.toFloat() / durationMs.toFloat()).coerceIn(0f, 0.99f)
                         onProgress(progress)
                     }
                 },
